Refactor: Decouple config fetching from usage
