+++ frontend/app/src/entities/authentication/ui/login.tsx
@@ -1,6 +1,6 @@
 import { LoginWithSSOButtons } from "@/entities/authentication/ui/login-sso-buttons";
 import { useAuth } from "@/entities/authentication/ui/useAuth";
-import { useConfig } from "@/entities/config/get-config.query";
+import { useConfig } from "@/entities/config/ui/config-provider";
 import { Button } from "@/shared/components/buttons/button-primitive";
 import InputField from "@/shared/components/form/fields/input.field";
 import PasswordInputField from "@/shared/components/form/fields/password-input.field";
@@ -10,7 +10,7 @@ import { classNames } from "@/shared/utils/common";
 import { useState } from "react";
 
 export const Login = () => {
-  const { data: config } = useConfig();
+  const config = useConfig();
   const [displaySSO, setDisplaySSO] = useState(true);
 
   if (config && config.sso.enabled && config.sso.providers && config.sso.providers.length > 0) {
