+++ frontend/app/src/entities/config/domain/get-config.ts
