+++ frontend/app/src/entities/config/ui/config-provider.tsx
@@ -1,0 +1,33 @@
+import { useGetConfig } from "@/entities/config/domain/get-config.query";
+import { ConfigAPI } from "@/entities/config/types";
+import ErrorScreen from "@/shared/components/errors/error-screen";
+import { InfrahubLoading } from "@/shared/components/loading/infrahub-loading";
+import React from "react";
+
+type ConfigContext = ConfigAPI;
+
+export const ConfigContext = React.createContext<ConfigContext>(null as unknown as ConfigContext);
+
+export function useConfig() {
+  const context = React.use(ConfigContext);
+
+  if (!context) {
+    throw new Error("useConfig must be used within a ConfigContext.");
+  }
+
+  return context;
+}
+
+export const ConfigProvider = ({ children }: { children: React.ReactNode }) => {
+  const { isPending, error, data } = useGetConfig();
+
+  if (isPending) {
+    return <InfrahubLoading>Loading config...</InfrahubLoading>;
+  }
+
+  if (error) {
+    return <ErrorScreen message={error.message} />;
+  }
+
+  return <ConfigContext value={data}>{children}</ConfigContext>;
+};
